+++ client/src/Components/DialogBox.js
@@ -2,7 +2,7 @@ import React, {useContext} from 'react'
 import { ExhibitionContext } from '../Context/Exhibition'
 import {ArtistContext} from '../Context/Artist'
 
-export default function DialogBox({artRef, message, handleDecline}) {
+export default function DialogBox({artRef, message, handleClearDialog}) {
 
     const {handleDeletedArtworkInExhibitionArray} = useContext(ExhibitionContext)
     const { handleDeletedArtworkInArtistList} = useContext(ArtistContext)
@@ -13,11 +13,12 @@ console.log( message)
 const handlePermDeleteArtwork = () => {
     const art = artRef.current;
 
-    fetch(`/artworks/${art.id}`, {
+    fetch(`/artworks/${art.id}?confirm_delete=true`, {
       method: 'DELETE'
     }).then(() => {
       handleDeletedArtworkInArtistList(art)
       handleDeletedArtworkInExhibitionArray(art)
+      handleClearDialog()
     })
   }
 
@@ -29,7 +30,7 @@ const handlePermDeleteArtwork = () => {
         <div>
           <h3>{message}</h3>
             <button className='dialogConfirmBtn' onClick={handlePermDeleteArtwork}>Delete Anyway</button>
-            <button className='dialogCancelBtn' onClick={handleDecline}>Cancel</button>
+            <button className='dialogCancelBtn' onClick={handleClearDialog}>Cancel</button>
         </div>
         </div>
